requirements.txt
# torch==1.13.1
# torchvision==0.14.1
# torchaudio==0.13.1
mmcv==1.6.0
matplotlib
argparse
pillow
numpy
# lpips
# plyfile
# pytorch_msssim
# open3d
# imageio[ffmpeg]

